Put material color utilities in a separate Gradle module
This avoids polluting our main sourceset and makes it easier to substitute for a proper Gradle dependency in the future
