+++ app/build.gradle
@@ -335,6 +335,7 @@ dependencies {
     gplayImplementation "com.google.firebase:firebase-messaging:23.0.7"
 
     // implementation 'androidx.activity:activity-ktx:1.4.0'
+    implementation project(':material-color-utilities')
 }
 
 task installGitHooks(type: Copy, group: "development") {
