+++ app/src/main/java/com/nextcloud/talk/ui/theme/ServerThemeImpl.kt
@@ -26,7 +26,7 @@ package com.nextcloud.talk.ui.theme
 import com.nextcloud.talk.R
 import com.nextcloud.talk.models.json.capabilities.ThemingCapability
 import com.nextcloud.talk.utils.ui.ColorUtil
-import scheme.Scheme
+import com.material_foundation.material_color_utilities.scheme.Scheme
 
 internal class ServerThemeImpl(themingCapability: ThemingCapability?, colorUtil: ColorUtil) :
     ServerTheme {
