+++ material-color-utilities/README.txt
@@ -1,0 +1,2 @@
+Imported from https://github.com/material-foundation/material-color-utilities with license APACHE-2.0,to be substituted
+with a Gradle dependency when they publish it.
