+++ material-color-utilities/src/main/java/com/material_foundation/material_color_utilities/blend/Blend.java
@@ -16,12 +16,12 @@
 
 // This file is automatically generated. Do not modify it.
 
-package blend;
+package com.material_foundation.material_color_utilities.blend;
 
-import hct.Cam16;
-import hct.Hct;
-import utils.ColorUtils;
-import utils.MathUtils;
+import com.material_foundation.material_color_utilities.hct.Cam16;
+import com.material_foundation.material_color_utilities.hct.Hct;
+import com.material_foundation.material_color_utilities.utils.ColorUtils;
+import com.material_foundation.material_color_utilities.utils.MathUtils;
 
 /** Functions for blending in HCT and CAM16. */
 public class Blend {
