+++ material-color-utilities/src/main/java/com/material_foundation/material_color_utilities/hct/HctSolver.java
@@ -16,10 +16,10 @@
 
 // This file is automatically generated. Do not modify it.
 
-package hct;
+package com.material_foundation.material_color_utilities.hct;
 
-import utils.ColorUtils;
-import utils.MathUtils;
+import com.material_foundation.material_color_utilities.utils.ColorUtils;
+import com.material_foundation.material_color_utilities.utils.MathUtils;
 
 /** A class that solves the HCT equation. */
 public class HctSolver {
