+++ material-color-utilities/src/main/java/com/material_foundation/material_color_utilities/hct/ViewingConditions.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package hct;
+package com.material_foundation.material_color_utilities.hct;
 
-import utils.ColorUtils;
-import utils.MathUtils;
+import com.material_foundation.material_color_utilities.utils.ColorUtils;
+import com.material_foundation.material_color_utilities.utils.MathUtils;
 
 /**
  * In traditional color spaces, a color can be identified solely by the observer's measurement of
