+++ material-color-utilities/src/main/java/com/material_foundation/material_color_utilities/quantize/QuantizerResult.java
@@ -14,21 +14,15 @@
  * limitations under the License.
  */
 
-package utils;
+package com.material_foundation.material_color_utilities.quantize;
 
-/** Utility methods for string representations of colors. */
-final class StringUtils {
-  private StringUtils() {}
+import java.util.Map;
 
-  /**
-   * Hex string representing color, ex. #ff0000 for red.
-   *
-   * @param argb ARGB representation of a color.
-   */
-  public static String hexFromArgb(int argb) {
-    int red = ColorUtils.redFromArgb(argb);
-    int blue = ColorUtils.blueFromArgb(argb);
-    int green = ColorUtils.greenFromArgb(argb);
-    return String.format("#%02x%02x%02x", red, green, blue);
+/** Represents result of a quantizer run */
+public final class QuantizerResult {
+  public final Map<Integer, Integer> colorToCount;
+
+  QuantizerResult(Map<Integer, Integer> colorToCount) {
+    this.colorToCount = colorToCount;
   }
 }
