+++ material-color-utilities/src/main/java/com/material_foundation/material_color_utilities/quantize/QuantizerWsmeans.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package quantize;
+package com.material_foundation.material_color_utilities.quantize;
 
 import static java.lang.Math.min;
