+++ material-color-utilities/src/main/java/com/material_foundation/material_color_utilities/scheme/Scheme.java
@@ -16,13 +16,13 @@
 
 // This file is automatically generated. Do not modify it.
 
-package scheme;
+package com.material_foundation.material_color_utilities.scheme;
 
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import com.google.errorprone.annotations.CheckReturnValue;
-import palettes.CorePalette;
+import com.material_foundation.material_color_utilities.palettes.CorePalette;
 
-/** Represents a Material color scheme, a mapping of color roles to colors. */
+/** Represents a Material color com.material_foundation.material_color_utilities.scheme, a mapping of color roles to colors. */
 @CheckReturnValue
 public class Scheme {
   private int primary;
