+++ material-color-utilities/src/main/java/com/material_foundation/material_color_utilities/utils/ColorUtils.java
@@ -16,7 +16,7 @@
 
 // This file is automatically generated. Do not modify it.
 
-package utils;
+package com.material_foundation.material_color_utilities.utils;
 
 /**
  * Color science utilities.
