+++ material-color-utilities/src/main/java/com/material_foundation/material_color_utilities/utils/MathUtils.java
@@ -16,7 +16,7 @@
 
 // This file is automatically generated. Do not modify it.
 
-package utils;
+package com.material_foundation.material_color_utilities.utils;
 
 /** Utility methods for mathematical operations. */
 public class MathUtils {
