+++ material-color-utilities/src/main/java/com/material_foundation/material_color_utilities/utils/StringUtils.java
@@ -14,8 +14,21 @@
  * limitations under the License.
  */
 
-package quantize;
+package com.material_foundation.material_color_utilities.utils;
 
-interface Quantizer {
-  public QuantizerResult quantize(int[] pixels, int maxColors);
+/** Utility methods for string representations of colors. */
+final class StringUtils {
+  private StringUtils() {}
+
+  /**
+   * Hex string representing color, ex. #ff0000 for red.
+   *
+   * @param argb ARGB representation of a color.
+   */
+  public static String hexFromArgb(int argb) {
+    int red = ColorUtils.redFromArgb(argb);
+    int blue = ColorUtils.blueFromArgb(argb);
+    int green = ColorUtils.greenFromArgb(argb);
+    return String.format("#%02x%02x%02x", red, green, blue);
+  }
 }
